+++ packages/opentelemetry-resources/test/Resource.test.ts
@@ -24,40 +24,36 @@ import {
 } from '@opentelemetry/semantic-conventions';
 import * as assert from 'assert';
 import * as sinon from 'sinon';
-import { Resource } from '../src';
 import { describeBrowser, describeNode } from './util';
+import {
+  DEFAULT_RESOURCE,
+  EMPTY_RESOURCE,
+  resourceFromAttributes,
+} from '../src';
 
 describe('Resource', () => {
-  const resource1 = new Resource({
-    attributes: {
+  const resource1 = resourceFromAttributes({
+    'k8s.io/container/name': 'c1',
+    'k8s.io/namespace/name': 'default',
+    'k8s.io/pod/name': 'pod-xyz-123',
+  });
+  const resource2 = resourceFromAttributes({
+    'k8s.io/zone': 'zone1',
+    'k8s.io/location': 'location',
+  });
+  const resource3 = resourceFromAttributes({
+    'k8s.io/container/name': 'c2',
+    'k8s.io/location': 'location1',
+  });
+  const emptyResource = resourceFromAttributes({});
+
+  it('should return merged resource', () => {
+    const expectedResource = resourceFromAttributes({
       'k8s.io/container/name': 'c1',
       'k8s.io/namespace/name': 'default',
       'k8s.io/pod/name': 'pod-xyz-123',
-    },
-  });
-  const resource2 = new Resource({
-    attributes: {
       'k8s.io/zone': 'zone1',
       'k8s.io/location': 'location',
-    },
-  });
-  const resource3 = new Resource({
-    attributes: {
-      'k8s.io/container/name': 'c2',
-      'k8s.io/location': 'location1',
-    },
-  });
-  const emptyResource = new Resource({ attributes: {} });
-
-  it('should return merged resource', () => {
-    const expectedResource = new Resource({
-      attributes: {
-        'k8s.io/container/name': 'c1',
-        'k8s.io/namespace/name': 'default',
-        'k8s.io/pod/name': 'pod-xyz-123',
-        'k8s.io/zone': 'zone1',
-        'k8s.io/location': 'location',
-      },
     });
     const actualResource = resource1.merge(resource2);
     assert.strictEqual(Object.keys(actualResource.attributes).length, 5);
@@ -68,13 +64,11 @@ describe('Resource', () => {
   });
 
   it('should return merged resource when collision in attributes', () => {
-    const expectedResource = new Resource({
-      attributes: {
-        'k8s.io/container/name': 'c2',
-        'k8s.io/namespace/name': 'default',
-        'k8s.io/pod/name': 'pod-xyz-123',
-        'k8s.io/location': 'location1',
-      },
+    const expectedResource = resourceFromAttributes({
+      'k8s.io/container/name': 'c2',
+      'k8s.io/namespace/name': 'default',
+      'k8s.io/pod/name': 'pod-xyz-123',
+      'k8s.io/location': 'location1',
     });
     const actualResource = resource1.merge(resource3);
     assert.strictEqual(Object.keys(actualResource.attributes).length, 4);
@@ -103,12 +97,10 @@ describe('Resource', () => {
   });
 
   it('should accept string, number, and boolean values', () => {
-    const resource = new Resource({
-      attributes: {
-        'custom.string': 'strvalue',
-        'custom.number': 42,
-        'custom.boolean': true,
-      },
+    const resource = resourceFromAttributes({
+      'custom.string': 'strvalue',
+      'custom.number': 42,
+      'custom.boolean': true,
     });
     assert.strictEqual(resource.attributes['custom.string'], 'strvalue');
     assert.strictEqual(resource.attributes['custom.number'], 42);
@@ -117,12 +109,10 @@ describe('Resource', () => {
 
   it('should log when accessing attributes before async attributes promise has settled', () => {
     const debugStub = sinon.spy(diag, 'error');
-    const resource = new Resource({
-      attributes: {
-        async: new Promise(resolve => {
-          setTimeout(resolve, 1);
-        }),
-      },
+    const resource = resourceFromAttributes({
+      async: new Promise(resolve => {
+        setTimeout(resolve, 1);
+      }),
     });
 
     resource.attributes;
@@ -140,29 +130,23 @@ describe('Resource', () => {
     });
 
     it('should return false for asyncAttributesPending if no promise provided', () => {
-      assert.ok(
-        !new Resource({ attributes: { foo: 'bar' } }).asyncAttributesPending
-      );
-      assert.ok(!Resource.EMPTY.asyncAttributesPending);
-      assert.ok(!Resource.default().asyncAttributesPending);
+      assert.ok(!resourceFromAttributes({ foo: 'bar' }).asyncAttributesPending);
+      assert.ok(!EMPTY_RESOURCE.asyncAttributesPending);
+      assert.ok(!DEFAULT_RESOURCE.asyncAttributesPending);
     });
 
     it('should return false for asyncAttributesPending once promise settles', async () => {
-      const resourceResolve = new Resource({
-        attributes: {
-          async: new Promise(resolve => {
-            setTimeout(resolve, 1);
-          }),
-        },
+      const resourceResolve = resourceFromAttributes({
+        async: new Promise(resolve => {
+          setTimeout(resolve, 1);
+        }),
       });
-      const resourceReject = new Resource({
-        attributes: {
-          async: new Promise((_, reject) => {
-            setTimeout(() => {
-              reject(new Error('reject'));
-            }, 1);
-          }),
-        },
+      const resourceReject = resourceFromAttributes({
+        async: new Promise((_, reject) => {
+          setTimeout(() => {
+            reject(new Error('reject'));
+          }, 1);
+        }),
       });
 
       for (const resource of [resourceResolve, resourceReject]) {
@@ -173,14 +157,12 @@ describe('Resource', () => {
     });
 
     it('should merge async attributes into sync attributes once resolved', async () => {
-      const resource = new Resource({
-        attributes: {
-          sync: 'fromsync',
-          // async attribute resolves after 1ms
-          async: new Promise(resolve =>
-            setTimeout(() => resolve('fromasync'), 1)
-          ),
-        },
+      const resource = resourceFromAttributes({
+        sync: 'fromsync',
+        // async attribute resolves after 1ms
+        async: new Promise(resolve =>
+          setTimeout(() => resolve('fromasync'), 1)
+        ),
       });
 
       await resource.waitForAsyncAttributes?.();
@@ -191,29 +173,21 @@ describe('Resource', () => {
     });
 
     it('should merge async attributes when both resources have promises', async () => {
-      const resource1 = new Resource({
-        attributes: {
-          promise1: Promise.resolve('promise1val'),
-          shared: Promise.resolve('promise1val'),
-        },
+      const resource1 = resourceFromAttributes({
+        promise1: Promise.resolve('promise1val'),
+        shared: Promise.resolve('promise1val'),
       });
-      const resource2 = new Resource({
-        attributes: {
-          promise2: Promise.resolve('promise2val'),
-          shared: Promise.resolve('promise2val'),
-        },
+      const resource2 = resourceFromAttributes({
+        promise2: Promise.resolve('promise2val'),
+        shared: Promise.resolve('promise2val'),
       });
       // this one rejects
-      const resource3 = new Resource({
-        attributes: {
-          err: Promise.reject(new Error('reject')),
-        },
+      const resource3 = resourceFromAttributes({
+        err: Promise.reject(new Error('reject')),
       });
-      const resource4 = new Resource({
-        attributes: {
-          promise4: Promise.resolve('promise4val'),
-          shared: Promise.resolve('promise4val'),
-        },
+      const resource4 = resourceFromAttributes({
+        promise4: Promise.resolve('promise4val'),
+        shared: Promise.resolve('promise4val'),
       });
 
       const merged = resource1
@@ -232,18 +206,14 @@ describe('Resource', () => {
     });
 
     it('should merge async attributes correctly when resource1 fulfils after resource2', async () => {
-      const resource1 = new Resource({
-        attributes: {
-          promise1: Promise.resolve('promise1val'),
-          shared: Promise.resolve('promise1val'),
-        },
+      const resource1 = resourceFromAttributes({
+        promise1: Promise.resolve('promise1val'),
+        shared: Promise.resolve('promise1val'),
       });
 
-      const resource2 = new Resource({
-        attributes: {
-          promise2: 'promise2val',
-          shared: 'promise2val',
-        },
+      const resource2 = resourceFromAttributes({
+        promise2: 'promise2val',
+        shared: 'promise2val',
       });
 
       const merged = resource1.merge(resource2);
@@ -258,18 +228,14 @@ describe('Resource', () => {
     });
 
     it('should merge async attributes correctly when resource2 fulfils after resource1', async () => {
-      const resource1 = new Resource({
-        attributes: {
-          promise1: Promise.resolve('promise1val'),
-          shared: 'promise1val',
-        },
+      const resource1 = resourceFromAttributes({
+        promise1: Promise.resolve('promise1val'),
+        shared: 'promise1val',
       });
 
-      const resource2 = new Resource({
-        attributes: {
-          promise2: new Promise(res => setTimeout(() => res('promise2val'), 1)),
-          shared: new Promise(res => setTimeout(() => res('promise2val'), 1)),
-        },
+      const resource2 = resourceFromAttributes({
+        promise2: new Promise(res => setTimeout(() => res('promise2val'), 1)),
+        shared: new Promise(res => setTimeout(() => res('promise2val'), 1)),
       });
 
       const merged = resource1.merge(resource2);
@@ -286,8 +252,8 @@ describe('Resource', () => {
     it('should log when promise rejects', async () => {
       const debugStub = sinon.spy(diag, 'debug');
 
-      const resource = new Resource({
-        attributes: { rejected: Promise.reject(new Error('rejected')) },
+      const resource = resourceFromAttributes({
+        rejected: Promise.reject(new Error('rejected')),
       });
       await resource.waitForAsyncAttributes?.();
 
@@ -301,7 +267,7 @@ describe('Resource', () => {
 
   describeNode('.default()', () => {
     it('should return a default resource', () => {
-      const resource = Resource.default();
+      const resource = DEFAULT_RESOURCE;
       assert.strictEqual(
         resource.attributes[SEMRESATTRS_TELEMETRY_SDK_NAME],
         SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_NAME]
@@ -323,7 +289,7 @@ describe('Resource', () => {
 
   describeBrowser('.default()', () => {
     it('should return a default resource', () => {
-      const resource = Resource.default();
+      const resource = DEFAULT_RESOURCE;
       assert.strictEqual(
         resource.attributes[SEMRESATTRS_TELEMETRY_SDK_NAME],
         SDK_INFO[SEMRESATTRS_TELEMETRY_SDK_NAME]