+++ CHANGELOG.md
@@ -80,6 +80,11 @@ For semantic convention package changes, see the [semconv CHANGELOG](packages/se
 * chore!: Raise the minimum supported Node.js version to `^18.19.0 || >=20.6.0`. Support for Node.js 14, 16, and early minor versions of 18 and 20 have been dropped. This applies to all packages except the 'api' and 'semantic-conventions' packages. [#5395](https://github.com/open-telemetry/opentelemetry-js/issues/5395) @trentm
 * feat(core)!: remove TracesSamplerValues from exports [#5406](https://github.com/open-telemetry/opentelemetry-js/pull/5406) @pichlermarc
   * (user-facing): TracesSamplerValues was only consumed internally and has been removed from exports without replacement
+* feat(resource)!: Remove resource class export in favor of functions and types only to aid in cross-version compatibility [#5421](https://github.com/open-telemetry/opentelemetry-js/pull/5421)
+  * Renames `Resource` class to `ResourceImpl` and makes it package-private
+  * Renames `IResource` interface to `Resource`
+  * Export function `resourceFromAttributes` to create a `Resource` from a `DetectedAttributes` object
+  * Only export types and functions. This aids in cross-version compatibility and makes it more easily extensible in the future.
 
 ### :rocket: (Enhancement)